move types to olli, make olli a peer dependency of olli-adapters
