+++ packages/adapters/package.json
@@ -30,6 +30,9 @@
     "vega-lite": "^5.3.0",
     "webpack": "^5.73.0"
   },
+  "peerDependencies": {
+    "olli": "^1.0.3"
+  },
   "devDependencies": {
     "webpack-cli": "^4.10.0"
   }
