+++ packages/core/src/Render/Table/index.ts
@@ -1,4 +1,4 @@
-import { OlliDatum } from "olli-adapters/src/Types";
+import { OlliDatum } from "../../Types";
 import { AccessibilityTreeNode } from "../../Structure/Types";
 
 /**
